+++ ainpla.f90
@@ -11,7 +11,7 @@ module ainpla_mod
 contains
 
   subroutine ainpla
-    !implicit integer (i-n), real*8 (a-h,o-z)
+    !implicit integer (i-n), real(c_double) (a-h,o-z)
     implicit none
     integer :: k
     integer :: k1
