+++ cfpgamma.f90
@@ -11,7 +11,7 @@ module cfpgamma_mod
 contains
 
   subroutine cfpgamma
-    !implicit integer (i-n), real*8 (a-h,o-z)
+    !implicit integer (i-n), real(c_double) (a-h,o-z)
     implicit none
     real(c_double) :: deby
     real(c_double) :: dense
