+++ eqfn.f90
@@ -11,10 +11,10 @@ module eqfn_mod
 
 contains
 
-      real*8 function eqfn(e,scalfct)
+      real(c_double) function eqfn(e,scalfct)
       use param_mod
       use comm_mod
-      implicit integer (i-n), real*8 (a-h,o-z)
+      implicit integer (i-n), real(c_double) (a-h,o-z)
 
 
 !..................................................................
