+++ equilib.f90
@@ -27,8 +27,8 @@ contains
   subroutine equilib(Requil,Zequil,index,PSIequil,BReq,BPHIeq,BZeq)
     use param_mod
     use comm_mod
-    implicit integer (i-n), real*8 (a-h,o-z)
-
+    implicit integer (i-n), real(c_double) (a-h,o-z)
+    
     !  Call with index=0 for setup,
     !            index=1 for PSI,BR,BPHI,BZ return (cgs units)
     !                    for a given Requil,Zequil coordinate.
