+++ fle.f90
@@ -18,7 +18,7 @@ contains
       use param_mod
       use comm_mod
       use r8subs_mod, only : luf
-      implicit integer (i-n), real*8 (a-h,o-z)
+      implicit integer (i-n), real(c_double) (a-h,o-z)
       character*(*) setup
 !
 !.......................................................................
@@ -223,7 +223,7 @@ contains
       use param_mod
       use comm_mod
       use r8subs_mod, only : luf
-      implicit integer (i-n), real*8 (a-h,o-z)
+      implicit integer (i-n), real(c_double) (a-h,o-z)
       character*(*) setup
 !
 !.......................................................................
@@ -386,7 +386,7 @@ contains
       use param_mod
       use comm_mod
       use r8subs_mod, only : dcopy
-      implicit integer (i-n), real*8 (a-h,o-z)
+      implicit integer (i-n), real(c_double) (a-h,o-z)
       character*(*) setup
 !
 !.......................................................................
