+++ frnfreya.f
@@ -4,13 +4,13 @@ c
      1     hibrz_,mfm1_,noplots)
       use param_mod
       use frplteq_mod, only :frplteq
-      implicit integer (i-n), real*8 (a-h,o-z)
+      implicit integer (i-n), real(c_double) (a-h,o-z)
 
 c.................................................................
       include 'frcomm.h77'
 c     ONETWO DIVERGENCE: SEE COMMENTS AT BEGINNING OF FREYA
       character*8 frmod_,fr_gyro_,beamplse_,noplots,codeid
-      real*8, intent(out) :: hibrz_(kz,ke,kb)
+      real(c_double), intent(out) :: hibrz_(kz,ke,kb)
       integer, intent(out) :: mfm1_
 
       character*8 ifirst
