+++ frplteq.f90
@@ -17,7 +17,7 @@ contains
     use param_mod
     use comm_mod
     use r8subs_mod, only : dcopy
-    implicit integer (i-n), real*8 (a-h,o-z)
+    implicit integer (i-n), real(c_double) (a-h,o-z)
     save
     !MPIINSERT_INCLUDE
 
@@ -297,7 +297,7 @@ contains
   end subroutine frplteq
 
   subroutine micfrplt
-    implicit integer (i-n), real*8 (a-h,o-z)
+    implicit integer (i-n), real(c_double) (a-h,o-z)
     save
 
     textt(1)="1$"
